+++ src/main/c/cext/ruby.c
@@ -12,6 +12,8 @@ void* rb_tr_cext;
 void* (*rb_tr_unwrap)(VALUE obj);
 VALUE (*rb_tr_wrap)(void *obj);
 VALUE (*rb_tr_longwrap)(long obj);
+VALUE (*rb_tr_id2sym)(ID id);
+ID (*rb_tr_sym2id)(VALUE val);
 
 void rb_tr_init_printf(void);
 void rb_tr_init_exception(void);
@@ -20,6 +22,8 @@ void rb_tr_init_exception(void);
 void rb_tr_init(void *ruby_cext) {
   rb_tr_cext = ruby_cext;
   rb_tr_unwrap = polyglot_invoke(rb_tr_cext, "rb_tr_unwrap_function");
+  rb_tr_id2sym = polyglot_invoke(rb_tr_cext, "rb_tr_id2sym_function");
+  rb_tr_sym2id = polyglot_invoke(rb_tr_cext, "rb_tr_sym2id_function");
   rb_tr_wrap = polyglot_invoke(rb_tr_cext, "rb_tr_wrap_function");
   rb_tr_longwrap = polyglot_invoke(rb_tr_cext, "rb_tr_wrap_function");